Rename chromeStateManager to chrome state
