+++ packages/chrome/src/ChromeProvider/index.ts
@@ -1,3 +1,3 @@
 export { default as ChromeProvider } from './ChromeProvider';
 export { default } from './ChromeProvider';
-export { ChromeContextState, UpdateEvents } from './chromeStateManager';
+export { ChromeContextState, UpdateEvents } from './chromeState';
